+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 0.4.9 (2024-07-18)
+
+### Features
+
+- added page locator and fix name of routes, app ([f7ef779](https://github.com/MarcinMG-web/react-dashboard/commit/f7ef779bc343421e853ac879ad4ef9b13b53bc47))
+- added pagination per pages ([#12](https://github.com/MarcinMG-web/react-dashboard/issues/12)) ([8e8c461](https://github.com/MarcinMG-web/react-dashboard/commit/8e8c4614278a00269c70383487404f963474ad07))
+
+### Bug Fixes
+
+- folder name ([c891c42](https://github.com/MarcinMG-web/react-dashboard/commit/c891c423f27043569971acf026e2d2a5543e15af))
+- folder name ([e3dbecf](https://github.com/MarcinMG-web/react-dashboard/commit/e3dbecf283ee878b3f4720abb1437f04d7571f36))
+- type and chip pending status ([4aa0f22](https://github.com/MarcinMG-web/react-dashboard/commit/4aa0f22df6a6a7138f08b73e398f8c6d04ec3ff3))
+
 ### 0.3.9 (2024-07-18)
 
 ### Features